scripts/requirements.txt
# Video AI Toolkit Dependencies
# Install: pip install -r requirements.txt

# API Clients
fal-client>=0.4.0
replicate>=0.20.0

# HTTP & Async
httpx>=0.27.0
aiohttp>=3.9.0
websockets>=12.0

# Image/Video Processing
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# CLI & Output
rich>=13.0.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0
